Allow qalloc_p to lower to dynamic num_wires in frontend
