+++ frontend/test/pytest/test_measurement_primitives.py
@@ -25,86 +25,6 @@ from catalyst.debug import get_compilation_stage, replace_ir
 from catalyst.jax_primitives import compbasis_p, expval_p, probs_p, state_p, var_p
 
 
-def test_dynamic_sample_backend_functionality():
-    """Test that a `sample` program with dynamic shots can be executed correctly."""
-
-    @catalyst.qjit(keep_intermediate=True)
-    def workflow_dyn_sample(shots):  # pylint: disable=unused-argument
-        # qml.device still needs concrete shots
-        device = qml.device("lightning.qubit", wires=1, shots=10)
-
-        @qml.qnode(device)
-        def circuit():
-            qml.RX(1.5, 0)
-            return qml.sample()
-
-        return circuit()
-
-    workflow_dyn_sample(10)
-    old_ir = get_compilation_stage(workflow_dyn_sample, "mlir")
-    workflow_dyn_sample.workspace.cleanup()
-
-    new_ir = old_ir.replace(
-        "catalyst.launch_kernel @module_circuit::@circuit() : () -> tensor<10x1xi64>",
-        "catalyst.launch_kernel @module_circuit::@circuit(%arg0) : (tensor<i64>) -> tensor<?x1xi64>",
-    )
-    new_ir = new_ir.replace(
-        "func.func public @circuit() -> tensor<10x1xi64>",
-        "func.func public @circuit(%arg0: tensor<i64>) -> tensor<?x1xi64>",
-    )
-    new_ir = new_ir.replace(
-        "quantum.device shots(%extracted) [",
-        """%shots = tensor.extract %arg0[] : tensor<i64>
-      quantum.device shots(%shots) [""",
-    )
-    new_ir = new_ir.replace("tensor<10x1x", "tensor<?x1x")
-
-    replace_ir(workflow_dyn_sample, "mlir", new_ir)
-    res = workflow_dyn_sample(37)
-    assert len(res) == 37
-
-    workflow_dyn_sample.workspace.cleanup()
-
-
-def test_dynamic_counts_backend_functionality():
-    """Test that a `counts` program with dynamic shots can be executed correctly."""
-
-    @catalyst.qjit(keep_intermediate=True)
-    def workflow_dyn_counts(shots):  # pylint: disable=unused-argument
-        # qml.device still needs concrete shots
-        device = qml.device("lightning.qubit", wires=1, shots=10)
-
-        @qml.qnode(device)
-        def circuit():
-            qml.RX(1.5, 0)
-            return qml.counts()
-
-        return circuit()
-
-    workflow_dyn_counts(10)
-    old_ir = get_compilation_stage(workflow_dyn_counts, "mlir")
-    workflow_dyn_counts.workspace.cleanup()
-
-    new_ir = old_ir.replace(
-        "catalyst.launch_kernel @module_circuit::@circuit() : () ->",
-        "catalyst.launch_kernel @module_circuit::@circuit(%arg0) : (tensor<i64>) ->",
-    )
-    new_ir = new_ir.replace(
-        "func.func public @circuit() ->", "func.func public @circuit(%arg0: tensor<i64>) ->"
-    )
-    new_ir = new_ir.replace(
-        "quantum.device shots(%extracted) [",
-        """%shots = tensor.extract %arg0[] : tensor<i64>
-      quantum.device shots(%shots) [""",
-    )
-
-    replace_ir(workflow_dyn_counts, "mlir", new_ir)
-    res = workflow_dyn_counts(4000)
-    assert res[1][0] + res[1][1] == 4000
-
-    workflow_dyn_counts.workspace.cleanup()
-
-
 def test_expval():
     """Test that the expval primitive can be captured by jaxpr."""
